Base app with select crop
Merged the base app gui with the area selection. Bug are still present.
Added undo and redo buttons along with the area selection button.

Predisposed the selection button to be moved out the editing bar and be placed with screeenshot button in order to immediately take a screen with the selection reducing so the numer of clicks required.

Co-Authored-By: Riccardo Drago <97728389+drgrico@users.noreply.github.com>